refactor: create scss and code cleanup
