+++ src/assets/base.scss
@@ -1,0 +1,4 @@
+:root {
+  --color-white: #ffffff;
+  --interactive-blue: #007AFF;
+}
