+++ src/main.ts
@@ -1,3 +1,4 @@
+import './assets/base.scss';
 import { createApp } from "vue";
 import { createPinia } from "pinia";
 import App from "./App.vue";
