+++ src/types/nav-items.types.ts
@@ -1,0 +1,4 @@
+export interface NavItems {
+  name: string;
+  displayText: string;
+}
